fix: 修复未赋初始值报错的问题
