+++ packages/taro-webpack-runner/src/util/chain.ts
@@ -7,12 +7,52 @@ import { partial } from 'lodash'
 import { mapKeys, pipe } from 'lodash/fp'
 import * as MiniCssExtractPlugin from 'mini-css-extract-plugin'
 import { join, resolve } from 'path'
+import { Bundler } from 'scss-bundle'
 import * as UglifyJsPlugin from 'uglifyjs-webpack-plugin'
 import * as webpack from 'webpack'
 
 import { recursiveMerge } from '.'
 import { getPostcssPlugins } from '../config/postcss.conf'
-import { CopyOptions, Option, PostcssOption } from './types'
+import {BuildConfig, CopyOptions, Option, PostcssOption} from './types'
+
+const makeConfig = async (config: BuildConfig) => {
+  const plugins = config.plugins || {}
+  const sassLoaderOption = config.sassLoaderOption || {}
+  const sass= plugins.sass || {}
+
+  let bundledContent = ''
+  if (sass.resource && sass.projectDirectory) {
+    const { resource, projectDirectory } = sass
+    const getBundleContent = async (url) => {
+      const bundler = new Bundler(undefined, projectDirectory)
+      const res = await bundler.Bundle(url)
+      bundledContent += res.bundledContent
+    }
+
+    try {
+      if (typeof resource === 'string') {
+        await getBundleContent(resource)
+      } else if (Array.isArray(resource)) {
+        for (const url of resource) {
+          await getBundleContent(url)
+        }
+      }
+    } catch (e) {
+      console.log(e)
+    }
+  }
+  if (sass.data) {
+    bundledContent += sass.data
+  }
+  return {
+    ...config,
+    plugins,
+    sassLoaderOption: {
+      ...sassLoaderOption,
+      data: sassLoaderOption.data ? `${sassLoaderOption.data}${bundledContent}` : bundledContent
+    }
+  }
+}
 
 const defaultUglifyJsOption = {
   keep_fnames: true,
@@ -261,7 +301,7 @@ const getModule = (appPath: string, {
   const extractCssLoader = getExtractCssLoader()
 
   const lastStyleLoader = enableExtract ? extractCssLoader : styleLoader
-  
+
   /**
    * css-loader 1.0.0版本移除了minimize选项...升级需谨慎
    *
@@ -421,7 +461,8 @@ const getDevtool = enableSourceMap => {
 export {
   isNodeModule,
   isTaroModule,
-  getEsnextModuleRules
+  getEsnextModuleRules,
+  makeConfig
 }
 
 export { getOutput, getMiniCssExtractPlugin, getHtmlWebpackPlugin, getDefinePlugin, processEnvOption, getHotModuleReplacementPlugin, getModule, getUglifyPlugin, getDevtool, getCssoWebpackPlugin, getCopyWebpackPlugin }
